% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compileRmd.R
\name{compileRmd}
\alias{compileRmd}
\title{Compiles RMarkdown file}
\usage{
compileRmd(file = NULL, format = "html", yaml = NULL, bibFile = NULL,
  cslFile = NULL, texEngine = NULL, bibEngine = NULL)
}
\arguments{
\item{file}{RMarkdown file.}

\item{format}{Output format. Choices are: 'md', 'html', 'htmlB', 'word', 'tex'.}

\item{yaml}{(Optional) Template file.
This is path to a .txt file containing YAML header.
If input Rmd file already contains YAML header, then set this to ''.
Otherwise, a default YAML header will be used depending on the output format.}

\item{bibFile}{(Optional) Bibtex file to use for citations.}

\item{cslFile}{(Optional) CSL file to use for citation styling.}

\item{texEngine}{(Optional) The latex command to use for tex output (e.g. 'pdflatex')}

\item{bibEngine}{(Optional) Choices are either 'biber' or 'bibtex'.}
}
\description{
Compiles Rmarkdown file to pdf, html, or word.
}

