% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getYamlHeader.R
\name{getYamlHeader}
\alias{getYamlHeader}
\title{Get Default YAML headers}
\usage{
getYamlHeader(format = NULL)
}
\arguments{
\item{format}{Output format. Choices are: 'html', 'htmlB', 'word', 'tex'.}
}
\description{
Returns default YAML headers depending on the format.
}

